Refactoring

+++ src/main/java/com/platzi/Market/persistence/services/ProductPersistenceService.java
@@ -1,15 +1,16 @@
 package com.platzi.Market.persistence.services;
 
 import java.util.List;
-import java.util.Optional;
 import org.springframework.stereotype.Service;
-import com.platzi.Market.domain.entities.ProductDomain;
+import com.platzi.Market.domain.dto.ProductSavedDTO;
+import com.platzi.Market.domain.dto.ProductReceivedDTO;
 import com.platzi.Market.exceptions.ValidationException;
-import com.platzi.Market.persistence.mappers.ProductMapper;
 import com.platzi.Market.persistence.entities.ProductEntity;
+import com.platzi.Market.persistence.entities.CategoryEntity;
 import org.springframework.beans.factory.annotation.Autowired;
+import com.platzi.Market.persistence.methods.DomainEntitiesMethods;
 import com.platzi.Market.persistence.repositories.ProductRepository;
-import com.platzi.Market.domain.entities.methods.DomainEntitiesMethods;
+import com.platzi.Market.persistence.repositories.CategoryRepository;
 
 @Service
 public class ProductPersistenceService implements DomainEntitiesMethods {
@@ -18,59 +19,83 @@ public class ProductPersistenceService implements DomainEntitiesMethods {
     private ProductRepository productRepository;
 
     @Autowired
-    private ProductMapper productMapper;
+    private CategoryRepository categoryRepository;
 
     // Para listar todos los productos.
     @Override
-    public List<ProductDomain> getAllProducts() {
+    public List<ProductSavedDTO> getAllProducts() {
         List<ProductEntity> productEntities = productRepository.findAll();
-        return productMapper.toProductsDomain(productEntities);
+        return productEntities.stream()
+                              .map(ProductSavedDTO::new)
+                              .toList();
     }
 
     // Para buscar productos por categoria.
     @Override
-    public Optional<List<ProductDomain>> getProductsByCategory(Long idCategory) {
-        if (!productRepository.existsByCategory(idCategory)) {
+    public List<ProductSavedDTO> getProductsByIdCategory(Long idCategory) {
+        if (!categoryRepository.existsByIdCategory(idCategory)) {
             throw new ValidationException("This category does not exist.");
         }
+        List<ProductEntity> productsEntityByCategoryList =
+                productRepository.findByCategory(idCategory);
 
-        List<ProductEntity> productsByCategory = productRepository.findByCategory(idCategory);
-        if (productsByCategory.isEmpty()) {
+        if (productsEntityByCategoryList.isEmpty()) {
             throw new ValidationException("The product category you are looking for is empty.");
         }
-        return Optional.of(productMapper.toProductsDomain(productsByCategory));
+        return productsEntityByCategoryList.stream()
+                                           .map(ProductSavedDTO::new)
+                                           .toList();
     }
 
     // Para buscar productos por la cantidad en stock
     @Override
-    public Optional<List<ProductDomain>> getProductsByStockQuantity(Long quantity) {
-        List<ProductEntity> productsStock = productRepository.findByStockQuantity(quantity);
-        if (productsStock.isEmpty()) {
+    public List<ProductSavedDTO> getProductsByStockQuantity(Long quantity) {
+        List<ProductEntity> productsStockList = productRepository.findByStockQuantity(quantity);
+
+        if (productsStockList.isEmpty()) {
             throw new ValidationException("No records were found with that quantity in stock.");
         }
-        return Optional.of(productMapper.toProductsDomain(productsStock));
+        return productsStockList.stream()
+                                .map(ProductSavedDTO::new)
+                                .toList();
     }
 
     // Para buscar un producto por su Id.
     @Override
-    public Optional<ProductDomain> getProductById(Long idProduct) {
+    public ProductSavedDTO getProductById(Long idProduct) {
         if (!productRepository.existsById(idProduct)) {
             throw new ValidationException("The product id '" + idProduct + "' not exists.");
         }
         ProductEntity productEntity = productRepository.getReferenceById(idProduct);
-        return Optional.of(productMapper.toProductDomain(productEntity));
+        return new ProductSavedDTO(productEntity);
     }
 
     // Para guardar un nuevo producto.
     @Override
-    public ProductDomain saveProduct(ProductDomain productDomain) {
-        ProductEntity productEntity = productMapper.toProductEntity(productDomain);
-        return productMapper.toProductDomain(productRepository.save(productEntity));
+    public ProductSavedDTO saveProduct(ProductReceivedDTO productReceivedDTO) {
+        if (!categoryRepository.existsById(productReceivedDTO.idCategory())) {
+            throw new ValidationException("This category does not exists.");
+        }
+        CategoryEntity categoryEntity =
+                categoryRepository.getReferenceById(productReceivedDTO.idCategory());
+
+        ProductEntity productEntity = new ProductEntity(
+                null,
+                productReceivedDTO.name(),
+                categoryEntity,
+                productReceivedDTO.barCode(),
+                productReceivedDTO.salesPrice(),
+                productReceivedDTO.stockQuantity()
+        );
+        return new ProductSavedDTO(productRepository.save(productEntity));
     }
 
     // Para eliminar un producto.
     @Override
     public void deleteProduct(Long idProduct) {
+        if (!productRepository.existsById(idProduct)) {
+            throw new ValidationException("This product cannot be removed because it does not exist.");
+        }
         productRepository.deleteById(idProduct);
     }
 }
